+++ core/tag/tags/MailTag.php
@@ -16,6 +16,12 @@ use luya\tag\BaseTag;
 class MailTag extends BaseTag
 {
     /**
+     * @var boolean Whether email obfuscation is enabled or not.
+     * @since 1.0.15
+     */
+    public $obfuscate = true;
+
+    /**
      * An example of how to use the MailTag.
      *
      * @return string The example string
@@ -49,6 +55,35 @@ EOT;
      */
     public function parse($value, $sub)
     {
-        return Html::mailto((!empty($sub)) ? $sub : $value, $value);
+        $label = $sub ?: $value;
+
+        // if obfuscation is enabled generate tag string due to yii tag method will encode attributes.
+        if ($this->obfuscate) {
+            if (!$sub) {
+                $label = $this->obfuscate($label);
+            }
+            return '<a href="'.$this->obfuscate("mailto:{$value}").'" rel="nofollow">'.$label.'</a>';
+        }
+
+        return Html::mailto($label, $value, [
+            'rel' => 'nofollow',
+        ]);
+    }
+
+    /**
+     * Obfucscate email adresse
+     *
+     * @param string $email
+     * @return string
+     * @see http://php.net/manual/de/function.bin2hex.php#11027
+     */
+    public function obfuscate($email)
+    {
+        $output = null;
+        for ($i = 0; $i < strlen($email); $i++) {
+            $output .= '&#'.ord($email[$i]).';';
+        }
+        
+        return $output;
     }
 }