+++ core/web/jsonld/CurrencyValue.php
@@ -5,9 +5,9 @@ namespace luya\web\jsonld;
 /**
  * The currency accepted.
  *
- * Use standard formats: ISO 4217 currency format e.g. "USD"; Ticker symbol for cryptocurrencies e.g. "BTC"; well known names 
+ * Use standard formats: ISO 4217 currency format e.g. "USD"; Ticker symbol for cryptocurrencies e.g. "BTC"; well known names
  * for Local Exchange Tradings Systems (LETS) and other currency types e.g. "Ithaca HOUR".
- * 
+ *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.14
  */
@@ -17,7 +17,7 @@ class CurrencyValue extends BaseValue
 
     public function __construct($currency)
     {
-        $this->_currency = $currency;    
+        $this->_currency = $currency;
     }
 
     public function getValue()