+++ core/web/jsonld/FoodEstablishmentTrait.php
@@ -4,7 +4,7 @@ namespace luya\web\jsonld;
 
 /**
  * Food Establishment Trait.
- * 
+ *
  * @author Basil Suter <basil@nadar.io>
  * @since 1.0.14
  */